requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gradio==4.7.1
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.24.3
sqlite3-python==0.1.0
pydantic==2.5.0
aiosqlite==0.19.0
python-multipart==0.0.6
